pg-listen-notify/requirements.txt
# psycopg2==2.7.7; sys_platform != 'win32' and python_version < '3.8'
# psycopg2==2.8.6; sys_platform == 'win32' or python_version >= '3.8'
psycopg2-binary